# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  http.CreateTransactionPayload:
    properties:
      items:
        items:
          properties:
            category:
              maximum: 3
              minimum: 1
              type: integer
            price:
              minimum: 0
              type: integer
            product_id:
              type: string
            product_name:
              type: string
            qty:
              minimum: 1
              type: integer
          required:
          - category
          - price
          - product_id
          - product_name
          - qty
          type: object
        minItems: 1
        type: array
      note:
        type: string
    required:
    - items
    type: object
...
  model.ProductCategory:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      _: Start with 0, but we'll ignore it
    x-enum-varnames:
    - _
    - Goods
    - Service
    - Subscription
  model.Transaction:
    properties:
      createdAt:
        type: string
      id:
        type: string
      invoiceCode:
        type: string
      is_paid:
        type: boolean
      note:
        type: string
      total:
        type: integer
      transactionDetails:
        items:
          $ref: '#/definitions/model.TransactionDetail'
        type: array
      updatedAt:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/model.User'
        description: Relationships
      userID:
        type: string
    type: object
  model.TransactionDetail:
    properties:
      category:
        $ref: '#/definitions/model.ProductCategory'
      createdAt:
        type: string
      id:
        type: string
      price:
        type: integer
      productID:
        type: string
      productName:
        type: string
      qty:
        type: integer
      transactionID:
        type: string
      updatedAt:
        type: string
    type: object
  model.User:
    properties:
      avatar_url:
        description: New field
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      image_status:
        description: New field
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  response.ApiResponse:
    properties:
      data: {}
      errors: {}
      meta:
        $ref: '#/definitions/response.Meta'
      status_code:
        type: integer
    type: object
  response.Meta:
    properties:
      current_page:
        type: integer
      per_page:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
host: localhost:3000
info:
  contact:
    email: sina4science@gmail.com
    name: Anis Fajar Fakhruddin
    url: https://discord.com/users/858389159555497994
  description: This is the API documentation for the Venturo Golang Core project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Venturo Golang Core API
  version: "1.0"
paths:
...
  /transactions:
    post:
      consumes:
      - application/json
      description: Creates a transaction with multiple detail items for the authenticated
        user.
      parameters:
      - description: Transaction Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.CreateTransactionPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created transaction
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Transaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new transaction
      tags:
      - Transactions
  /transactions/{id}/pay:
    post:
      description: Marks a transaction as paid and triggers a background report update.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully paid
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Transaction not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Pay for a Transaction
      tags:
      - Transactions
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and a JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"